+++ src/internal/domain/CSSRuleWrapper.ts
@@ -1,4 +1,4 @@
-import { PostCSSRule } from "../../Types";
+import { Rule } from "postcss";
 
 import {
     cleanSelector,
@@ -9,7 +9,7 @@ import {
 } from "./CSSSelector";
 
 export const prefixWrapCSSRule = (
-    cssRule: PostCSSRule,
+    cssRule: Rule,
     nested: string | null,
     ignoredSelectors: (string | RegExp)[],
     prefixSelector: string,
@@ -17,15 +17,13 @@ export const prefixWrapCSSRule = (
 ): void => {
     // Check each rule to see if it exactly matches our prefix selector, when
     // this happens, don't try to prefix that selector.
-    const rules = cssRule.selector
-        .split(",")
-        .filter(
-            (selector) =>
-                !cssRuleMatchesPrefixSelector(
-                    { selector: selector },
-                    prefixSelector,
-                ),
-        );
+    const rules = cssRule.selectors.filter(
+        (selector) =>
+            !cssRuleMatchesPrefixSelector(
+                { selector: selector },
+                prefixSelector,
+            ),
+    );
 
     if (rules.length === 0) {
         return;
@@ -48,7 +46,7 @@ export const prefixWrapCSSRule = (
 
 export const prefixWrapCSSSelector = (
     cssSelector: string,
-    cssRule: PostCSSRule,
+    cssRule: Rule,
     nested: string | null,
     ignoredSelectors: (string | RegExp)[],
     prefixSelector: string,